LPS-136143 Convert class to abstract since it is not necessary to create an instance of TemplateDisplayContext
